% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zarr.R
\name{pluot_register_store}
\alias{pluot_register_store}
\title{Register a pizzarr store for use with pluot renderers.}
\usage{
pluot_register_store(name, store)
}
\arguments{
\item{name}{Character string; the store name referenced in layer parameters.}

\item{store}{A pizzarr store object (e.g. MemoryStore, DirectoryStore).}
}
\description{
Register a pizzarr store for use with pluot renderers.
}
